+++ lib/adapters/openai_adapter.dart
@@ -1,6 +1,5 @@
 import 'package:dio/dio.dart';
 
-import '../core/ai_key_pool.dart';
 import '../core/ai_provider_adapter.dart';
 import '../core/retry_interceptor.dart';
 import '../models/ai_embedding.dart';