+++ test/adapters/openrouter_adapter_test.dart
@@ -1,7 +1,6 @@
 import 'dart:convert';
 
 import 'package:ai_core/ai_core.dart';
-import 'package:ai_core/models/openrouter_options.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter_test/flutter_test.dart';