+++ test/models/ai_response_format_test.dart
@@ -1,5 +1,3 @@
-import 'dart:convert';
-
 import 'package:ai_core/ai_core.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter_test/flutter_test.dart';